archive/legacy_iris_gateway_sep2024/requirements.txt
httpx>=0.27
python-dotenv>=1.0
typer>=0.12
rich>=13.7
orjson>=3.10